requirements.txt
certifi==2022.9.24
Faker==15.3.1
geopy==2.2.0
pymongo==4.3.2
python-dotenv==0.21.0
